+++ train_discriminator.py
@@ -159,7 +159,7 @@ def main(cfg: DictConfig):
                     param,
                     round(getattr(disc_trainer.attack, param), 4)
                 )
-    else:
+    else: 
         const_params["attack_name"] = cfg["attack"]["name"]
         trainer_params = dict(cfg["training_params"])
         trainer_params.update(const_params)
@@ -188,7 +188,10 @@ def main(cfg: DictConfig):
         logger = SummaryWriter(cfg["save_path"] + "/tensorboard")
 
     disc_trainer.train_model(train_loader, test_loader, augmentator, logger)
-    os.remove(attack_model_path)
+    if cfg['pretrained']:
+        os.remove(attack_model_path)
+    else:
+        pass
 
 
     if not cfg["test_run"]:
